Specify maximum number of function evaluations
The fit with the `lm` method had a fixed maximum number of 500
iterations. This is set in the `matrixfit` function. When looking at the
results of the fits one sometimes finds that the number of function
evaluations has been exhausted and not the number of iterations. Since
out function is cheap to evaluate and we would like to limit the number
of iterations, this change will specify a sufficiently large number of
function evaluations such that this will never be the limiting factor.