Pass the bootstrap id to the fit functions
This is a solution for Issue #97.

In the implementation there is one potential stumbling block: The
ellipsis `...` is used in the signature of `bootstrap.nlsfit` to capture
arbitrary auxiliary fixed arguments that will be passed to the model
function. Before the `...` were passed directly from `bootstrap.nlsfit`
to the function, now it is passed to `nls.lm`/`optim`, then to `wrapper`
and then to the actual model function. So basically the scope of the
`...` have changed and include the `boot_r` now.