Clean up C code to not write on stdout and stderr
C code used in R is not supposed to write to `stdout` and `stderr`
directly via `printf` and `fprinftf(stderr, …`). Instead one shall use
`Rprintf` and `REprintf`. Even better is to use `warning` such that the
output is added to the list of warnings.

Also I removed a bunch of headers and sorted the includes in the
standard way.